LayoutTests/imported/w3c/web-platform-tests/service-workers/stub-4.6.3-cache-storage-expected.txt

CONSOLE MESSAGE: line 251: callback not yet supported

PASS CacheStorage interface: existence and properties of interface object 
FAIL CacheStorage interface object length assert_equals: wrong value for CacheStorage.length expected 1 but got 0
PASS CacheStorage interface object name 
PASS CacheStorage interface: existence and properties of interface prototype object 
PASS CacheStorage interface: existence and properties of interface prototype object's "constructor" property 
PASS CacheStorage interface: operation match(ScalarValueString,DOMString) 
FAIL CacheStorage interface: operation get(DOMString) assert_own_property: interface prototype object missing non-static operation expected property "get" missing
PASS CacheStorage interface: operation has(DOMString) 
FAIL CacheStorage interface: operation set(DOMString,Cache) assert_own_property: interface prototype object missing non-static operation expected property "set" missing
FAIL CacheStorage interface: operation clear() assert_own_property: interface prototype object missing non-static operation expected property "clear" missing
PASS CacheStorage interface: operation delete(DOMString) 
FAIL CacheStorage interface: operation forEach(CacheStorageIterationCallback,object) assert_own_property: interface prototype object missing non-static operation expected property "forEach" missing
FAIL CacheStorage interface: operation entries() assert_own_property: interface prototype object missing non-static operation expected property "entries" missing
PASS CacheStorage interface: operation keys() 
FAIL CacheStorage interface: operation values() assert_own_property: interface prototype object missing non-static operation expected property "values" missing
FAIL CacheStorage interface: operation size() assert_own_property: interface prototype object missing non-static operation expected property "size" missing
FAIL CacheStorage interface: throw new Error ('No object defined for the CacheStorage interface') must inherit property "match" with the proper type (0) assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the CacheStorage interface"
FAIL CacheStorage interface: calling match(ScalarValueString,DOMString) on throw new Error ('No object defined for the CacheStorage interface') with too few arguments must throw TypeError assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the CacheStorage interface"
FAIL CacheStorage interface: throw new Error ('No object defined for the CacheStorage interface') must inherit property "get" with the proper type (1) assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the CacheStorage interface"
FAIL CacheStorage interface: calling get(DOMString) on throw new Error ('No object defined for the CacheStorage interface') with too few arguments must throw TypeError assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the CacheStorage interface"
FAIL CacheStorage interface: throw new Error ('No object defined for the CacheStorage interface') must inherit property "has" with the proper type (2) assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the CacheStorage interface"
FAIL CacheStorage interface: calling has(DOMString) on throw new Error ('No object defined for the CacheStorage interface') with too few arguments must throw TypeError assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the CacheStorage interface"
FAIL CacheStorage interface: throw new Error ('No object defined for the CacheStorage interface') must inherit property "set" with the proper type (3) assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the CacheStorage interface"
FAIL CacheStorage interface: calling set(DOMString,Cache) on throw new Error ('No object defined for the CacheStorage interface') with too few arguments must throw TypeError assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the CacheStorage interface"
FAIL CacheStorage interface: throw new Error ('No object defined for the CacheStorage interface') must inherit property "clear" with the proper type (4) assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the CacheStorage interface"
FAIL CacheStorage interface: throw new Error ('No object defined for the CacheStorage interface') must inherit property "delete" with the proper type (5) assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the CacheStorage interface"
FAIL CacheStorage interface: calling delete(DOMString) on throw new Error ('No object defined for the CacheStorage interface') with too few arguments must throw TypeError assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the CacheStorage interface"
FAIL CacheStorage interface: throw new Error ('No object defined for the CacheStorage interface') must inherit property "forEach" with the proper type (6) assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the CacheStorage interface"
FAIL CacheStorage interface: calling forEach(CacheStorageIterationCallback,object) on throw new Error ('No object defined for the CacheStorage interface') with too few arguments must throw TypeError assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the CacheStorage interface"
FAIL CacheStorage interface: throw new Error ('No object defined for the CacheStorage interface') must inherit property "entries" with the proper type (7) assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the CacheStorage interface"
FAIL CacheStorage interface: throw new Error ('No object defined for the CacheStorage interface') must inherit property "keys" with the proper type (8) assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the CacheStorage interface"
FAIL CacheStorage interface: throw new Error ('No object defined for the CacheStorage interface') must inherit property "values" with the proper type (9) assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the CacheStorage interface"
FAIL CacheStorage interface: throw new Error ('No object defined for the CacheStorage interface') must inherit property "size" with the proper type (10) assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the CacheStorage interface"
